Improve undeploy logic

+++ scripts/undeploy-olm.sh
@@ -40,14 +40,31 @@ do
     esac
 done
 
+set -o xtrace
+
+# We want to continue on even if we hit errors
+set +o errexit
+set +o pipefail
+
+# For debug purposes, show all the ClusterServiceVersions installed in the cluster
+kubectl get csv -A
+
 # The CSV is available in every namespace. We need to extract from it the
 # namespace where the operator pod is running.
 NAMESPACE=$(kubectl get csv -A -o jsonpath='{range .items[*]}{.metadata.annotations.olm\.operatorNamespace}{"\n"}{end}' | grep -v olm | head -1)
+if [[ -n $NAMESPACE ]]
+then
+    kubectl delete -n $NAMESPACE clusterserviceversion --selector operators.coreos.com/verticadb-operator.$NAMESPACE="" || :
+fi
 
-echo "Namespace: $NAMESPACE"
+subs=$(kubectl get subscription --all-namespaces -o jsonpath='{range .items[?(@.metadata.name=="e2e-verticadb-subscription")]}{.metadata.name}{" -n " }{.metadata.namespace}{"\n"}{end}')
+if [[ -n $subs ]]
+then
+    kubectl delete subscriptions $subs
+fi
 
-set -o xtrace
-
-kubectl delete -n $NAMESPACE clusterserviceversion --selector operators.coreos.com/verticadb-operator.$NAMESPACE="" || :
-kubectl delete -n $NAMESPACE operatorgroup e2e-operatorgroup || :
-kubectl delete -n $NAMESPACE subscription e2e-verticadb-subscription || :
+ogs=$(kubectl get operatorgroup --all-namespaces -o jsonpath='{range .items[?(@.metadata.name=="e2e-operatorgroup")]}{.metadata.name}{" -n " }{.metadata.namespace}{"\n"}{end}')
+if [[ -n $ogs ]]
+then
+    kubectl delete operatorgroups $ogs
+fi
